pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "gy-doc-search"
version = "0.1.0"
description = "Semantic search over project documentation for AI coding agents"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "OpenAI Codex" },
]
dependencies = [
  "click>=8.0",
  "jinja2>=3.0",
  "pyyaml>=6.0",
  "rich>=13.0",
]

[project.optional-dependencies]
chromadb = ["chromadb>=0.5.0"]
mcp = ["mcp[server]>=1.0.0"]
openai = ["openai>=1.0"]
search = ["rank-bm25>=0.2.2"]
sentence-transformers = ["sentence-transformers>=3.0.0"]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
]

[project.scripts]
gy-doc-search = "gy_doc_search.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
